+++ lua/plugins/completion.lua
@@ -1,0 +1,20 @@
+return {
+  "saghen/blink.cmp",
+  version = "1.*",
+  event = "InsertEnter",
+  opts = {
+    keymap = { preset = "default" },
+    completion = {
+      documentation = {
+        auto_show = true,
+        auto_show_delay_ms = 500,
+      },
+    },
+    sources = {
+      default = { "lsp", "path", "snippets", "buffer" },
+    },
+    signature = { enabled = false },
+    fuzzy = { implementation = "prefer_rust_with_warning" },
+  },
+  opts_extend = { "sources.default" },
+}