+++ lua/plugins/telescope.lua
@@ -1,0 +1,77 @@
+local function find_command()
+  if vim.fn.executable("fd") == 1 then
+    return { "fd", "--type", "f", "--hidden", "--exclude", ".git" }
+  end
+
+  if vim.fn.executable("fdfind") == 1 then
+    return { "fdfind", "--type", "f", "--hidden", "--exclude", ".git" }
+  end
+
+  return { "rg", "--files", "--hidden", "--glob", "!.git/*" }
+end
+
+return {
+  "nvim-telescope/telescope.nvim",
+  version = "*",
+  cmd = "Telescope",
+  dependencies = {
+    "nvim-lua/plenary.nvim",
+    {
+      "nvim-telescope/telescope-fzf-native.nvim",
+      build = "make",
+    },
+  },
+  keys = {
+    {
+      "<leader>ff",
+      function()
+        require("telescope.builtin").find_files({
+          hidden = true,
+          find_command = find_command(),
+        })
+      end,
+      desc = "Find files",
+    },
+    {
+      "<leader>fg",
+      function()
+        require("telescope.builtin").live_grep()
+      end,
+      desc = "Live grep",
+    },
+    {
+      "<leader>fb",
+      function()
+        require("telescope.builtin").buffers()
+      end,
+      desc = "Find buffers",
+    },
+    {
+      "<leader>fr",
+      function()
+        require("telescope.builtin").oldfiles()
+      end,
+      desc = "Recent files",
+    },
+    {
+      "<leader>fh",
+      function()
+        require("telescope.builtin").help_tags()
+      end,
+      desc = "Help tags",
+    },
+    {
+      "<leader>fd",
+      function()
+        require("telescope.builtin").diagnostics()
+      end,
+      desc = "Find diagnostics",
+    },
+  },
+  opts = {},
+  config = function(_, opts)
+    local telescope = require("telescope")
+    telescope.setup(opts)
+    pcall(telescope.load_extension, "fzf")
+  end,
+}